+++ packages/utils/src/index.ts
@@ -1,7 +1,6 @@
 export { callAll } from './utils/CallAll';
 export { throttle } from './utils/throttle';
-export { debounceThrottle } from './utils/debounceThrottle';
-export { debounce } from './utils/debounce';
+export * from './utils/debounce';
 export * as dateFormatters from './utils/dates/dateFormatters';
 export { mapObjIndexed } from './utils/objects/mapObjIndexed';
 export { generateRange } from './utils/pagination';
