--- packages/utils/src/utils/debounceThrottle/debounceThrottle.ts
@@ -1,24 +1,0 @@
-type UnknownFn = (...args: any[]) => unknown;
-
-export const debounceThrottle = (func: UnknownFn, wait = 200) => {
-  let timeoutId: NodeJS.Timeout | null = null;
-  const executedFunction = (...args: any[]) => {
-    const postponedFn = () => {
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-      func(...args);
-    };
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-    }
-    timeoutId = setTimeout(postponedFn, wait);
-  };
-
-  const cancel = function () {
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-    }
-  };
-  return [executedFunction, cancel];
-};
