--- packages/utils/src/utils/debounceThrottle/index.ts
@@ -1,1 +1,0 @@
-export { debounceThrottle } from './debounceThrottle';
