+++ packages/widgets/src/components/SearchBox/index.ts
@@ -2,4 +2,3 @@ export { SearchBox } from './SearchBox';
 export { SearchBoxCrossIcon } from './SearchBoxCrossIcon';
 export { SearchBoxInput } from './SearchBoxInput';
 export { SearchBoxWrapper } from './SearchBoxWrapper';
-export { StoryComponent } from './stories/StoryComponent';
