+++ packages/widgets/src/components/SearchBox/stories/StoryComponent.tsx
@@ -2,10 +2,14 @@ import { FieldValues, useForm } from 'react-hook-form';
 
 import { SearchBox } from '..';
 
-export const StoryComponent = () => {
+export const StoryComponent = ({
+  handleCallback,
+}: {
+  handleCallback?: (searchTerm: string) => void;
+}) => {
   const { register, control, resetField } = useForm<FieldValues>();
 
-  const handleCallback = (searchTerm: string) => {
+  const onCallback = (searchTerm: string) => {
     console.log('Searching for the term...', searchTerm);
   };
 
@@ -14,7 +18,7 @@ export const StoryComponent = () => {
       register={register}
       control={control}
       resetField={resetField}
-      callback={handleCallback}
+      callback={handleCallback || onCallback}
       name={'search'}
     />
   );
